requirements.txt
langchain
langchain-community
langchain-openai
langchain-experimental
python-dotenv
langchain_ollama
reportlab
python-docx
markdown
dotenv
fastapi
uvicorn[standard]
pydantic
sqlalchemy
alembic
python-multipart
# 认证相关依赖
python-jose[cryptography]
passlib[bcrypt]
email-validator